+++ src/entity.rs
@@ -8,10 +8,6 @@ use ggez::GameResult;
 
 const DEBUG: bool = false;
 
-const GRAVITY: f32 = 0.25;
-const JUMP_IMPULSE: f32 = 2.75;
-pub const SCREEN_TOP: f32 = -16.0;
-
 #[derive(Debug, PartialEq, Eq, Clone)]
 /// The current state of the game.
 pub enum PlayState {
@@ -23,12 +19,12 @@ pub enum PlayState {
 /// The physics on the entity.
 pub struct Physics {
     pub velocity: Vector2<f32>,
-    acceleration: Vector2<f32>,
-    gravity: bool,
+    pub acceleration: Vector2<f32>,
+    pub gravity: bool,
 }
 
 impl Physics {
-    fn new (with_gravity: bool) -> Self {
+    pub fn new (with_gravity: bool) -> Self {
         Self {
             velocity: Vector2::new(0.0, 0.0),
             acceleration: Vector2::new(0.0, 0.0),
@@ -47,156 +43,11 @@ pub enum ScoringPipe {
     Scored,
 }
 
-pub struct TileEntity {
-    pub sprite: Sprite,
-    pub position: Point2<f32>,
-}
-
-impl TileEntity {
-    pub fn draw(&mut self, ctx: &mut Context, batch: &mut SpriteBatch) {
-        let s = &mut self.sprite;
-        batch.add(s.add_draw_param(self.position.clone()));
-    }
-
-    pub fn new(sprite: Sprite, position: (f32, f32)) -> Self {
-        Self {
-            sprite,
-            position: Point2::new(position.0, position.1),
-        }
-    }
-}
-
 pub struct PipeEntity {
     pub sprite: Sprite,
     pub position: Point2<f32>,
     scroller: Option<Scroll>,
     pub scoring_pipe: Option<ScoringPipe>,
-}
-
-pub struct PlayerEntity {
-    pub sprite: Sprite,
-    pub position: Point2<f32>,
-    pub player_sprites: Vec<Sprite>,
-    can_jump: bool,
-    pub physics: Physics,
-}
-
-impl PlayerEntity {
-    pub fn update(
-        &mut self,
-        ctx: &mut Context,
-        state: &PlayState,
-    ) -> PlayState {
-        let physics = &mut self.physics;
-        physics.acceleration = if physics.gravity {
-            Vector2::new(0.0, GRAVITY)
-        } else {
-            Vector2::new(0.0, 0.0)
-        };
-
-
-        let mut state = state.clone();
-        if state.is_not_dead()
-        {
-            use ggez::event::KeyCode;
-            use ggez::input::keyboard;
-            if !keyboard::pressed_keys(ctx).contains(&KeyCode::Space) && !self.can_jump {
-                self.can_jump = true;
-            }
-
-            if keyboard::is_key_pressed(ctx, KeyCode::Space) && self.can_jump {
-                let physics = &mut self.physics;
-                PlayerEntity::jump(physics);
-
-                // exit start screen state.
-                if state == PlayState::StartScreen {
-                    state = PlayState::Play;
-                }
-            }
-        }
-
-        // Self jumping script on the start screen.
-        if state == PlayState::StartScreen {
-            self.auto_jump()
-        }
-
-        self.change_position(ctx);
-
-        // clamp y to not go above the top of the screen easily.
-        self.prevent_going_off();
-        state
-    }
-
-    fn change_position(&mut self, ctx: &mut Context) {
-        let delta = ggez::timer::delta(ctx).as_nanos() as f32;
-        let physics = &mut self.physics;
-        physics.acceleration.scale(1.0 / delta);
-        physics.velocity += physics.acceleration;
-        physics.velocity.scale(1.0 / delta);
-        // moves all the entities on the board.
-        self.position += physics.velocity;
-    }
-
-    pub fn new(sprite: Sprite, position: (f32, f32), player_sprites: Vec<Sprite>) -> Self {
-        Self {
-            sprite,
-            position: Point2::new(position.0, position.1),
-            physics: Physics::new(true),
-            can_jump: true,
-            player_sprites,
-        }
-    }
-    pub fn overlaps(&self, other : &PipeEntity) -> bool {
-        let player_rect = self.get_bounds();
-        let other_rect = other.get_bounds();
-
-        other_rect.overlaps(&player_rect)
-    }
-    pub fn get_bounds(&self) -> graphics::Rect {
-        let mut rect = self.sprite.get_bound_box();
-        rect.move_to(self.position.clone());
-
-        rect
-    }
-    fn prevent_going_off(&mut self) -> () {
-        self.position.y = if self.position.y < SCREEN_TOP {
-            SCREEN_TOP
-        } else {
-            self.position.y
-        }
-    }
-    fn auto_jump(&mut self) -> () {
-        let physics = &mut self.physics;
-        if self.position.y > 600.0 / 8.0 {
-            PlayerEntity::jump(physics);
-        }
-    }
-    pub fn draw(&mut self, batch: &mut SpriteBatch) -> GameResult {
-        self.draw_player(batch);
-
-        Ok(())
-    }
-    fn draw_player(&mut self, batch: &mut SpriteBatch) {
-        let s = &mut self.player_sprites;
-        let p = &self.physics;
-        // need velocity to map to these rotations between -0.2 and 0.2!
-        let angle = rescale_range(p.velocity.y, -7.0, 7.0, -0.6, 0.6);
-        let x = if p.velocity.y >= 0.0 {
-            &mut s[1]
-        } else {
-            &mut s[0]
-        };
-        batch.add(
-            x.add_draw_param(self.position.clone())
-                .offset(Point2::new(0.5, 0.5))
-                .rotation(angle),
-        );
-    }
-
-    fn jump(physics: &mut Physics) {
-        physics.acceleration = Vector2::new(0.0, -GRAVITY);
-        physics.velocity = Vector2::new(0.0, -JUMP_IMPULSE);
-    }
 }
 
 /// Everything that can be interacted with is an entity.
@@ -278,7 +129,7 @@ impl PipeEntity {
 
         // Moves the pipes towards the crab !
         let speed = pipe_velocity();
-        self.position += Vector2::new(speed, 0);
+        self.position += Vector2::new(speed, 0.0);
         // when the pipes go off the left side.
         // we put them back at right side to come again.
         self.recycle_passed_pipes(pipe_tracker);
@@ -326,11 +177,3 @@ impl PipeEntity {
     }
 }
 
-/// Returns an f32 scaled [oldMin, oldMax] into the range [newMin, newMax]
-/// Thanks https://stackoverflow.com/a/5295202/6421793
-fn rescale_range(value: f32, old_min: f32, old_max: f32, new_min: f32, new_max: f32) -> f32 {
-    use ggez::nalgebra::clamp;
-    let old_range = old_max - old_min;
-    let new_range = new_max - new_min;
-    (((clamp(value, old_min, old_max) - old_min) * new_range) / old_range) + new_min
-}
